requirements.txt
# Core dependencies
pandas>=2.0.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.0
aiofiles>=24.0.0

# Twitter scraping
twikit>=2.3.0

# OpenSea API & web requests  
requests>=2.32.0
aiohttp>=3.9.0

# Advanced sentiment analysis (Flare AI Consensus Learning)
structlog>=25.0.0
pydantic-settings>=2.9.0
anthropic>=0.52.0
openai>=1.82.0

# Data processing
numpy>=2.0.0
beautifulsoup4>=4.13.0 